Rain Particles and Main Menu backdrop
Added Rain Particles, started on the Main Menu back-drop environments, and the block-outs. Also, have the issue with the Starter Content stuff, so that is why there are so many changes. (When the merge happens, the buildmaster can look through the delete stuff we don't need, but I've moved over the fire and smoke, and whatnot to my folders)