Sounds and Errors and Locations sorted
Added sounds to buttons and sliders, sorted out an issue with the Audio menu that would not save values after leaving, sorted out sounds and organised them accordingly. Sorted out errors with the Main Menu animation. And lastly, made the third (and possibly final) location for the Main Menu backdrop
All that needs doing are some animations. After that, everything that is high priority is completely done, and I will focus on the Level Up menu.